First cut. WIP.
